Support using embedded test index for UI
See https://github.com/hbz/lobid-resources/issues/286
