+++ web/conf/resources.conf
@@ -12,7 +12,8 @@ index = {
 	cluster = {
 		name = "weywot"
 		#hosts = ["weywot4.hbz-nrw.de", "weywot5.hbz-nrw.de"]
-		hosts = ["10.9.0.13", "10.9.0.14"]
+		#hosts = ["10.9.0.13", "10.9.0.14"]
+		hosts = [] # use local index created from local test data
 		port = 9300
 	}
 }
